moved handshake-interface units to handshake folder
