+++ experimental/tools/unit-generators/vhdl/generators/handshake/cond_br.py
@@ -1,4 +1,4 @@
-from generators.support.join import generate_join
+from generators.handshake.join import generate_join
 
 
 def generate_cond_br(name, params):
