+++ experimental/tools/unit-generators/vhdl/generators/handshake/merge_notehb.py
